Improve all aspects.
Reworked package building process from `PKGBUILD` as a primary source of
information + `build-deb` sourcing it, to `package-builder`.

`package-builder` downloads are sources listed in `sources` file, and
instructions in `build` to build package to a directory. From this
directory either Debian, Arch Linux or tar package can be created.

The biggest step is building `luatex` manually. This additionally requires
bootstrapping `ctangle` as that is required for `mplib`.